+++ importanize/main.py
@@ -12,7 +12,11 @@ import six
 
 from . import __version__
 from .groups import ImportGroups
-from .parser import find_imports_from_lines, get_artifacts, parse_statements
+from .parser import (
+    find_imports_from_lines,
+    get_file_artifacts,
+    parse_statements,
+)
 from .utils import read
 
 
@@ -73,16 +77,19 @@ parser.add_argument(
     nargs='*',
     default=['.'],
     help='Path either to a file or directory where '
-         'all Python import will be organized. ',
+         'all Python files imports will be organized.',
 )
 parser.add_argument(
     '-c', '--config',
     type=six.text_type,
     default=default_config,
     help='Path to importanize config json file. '
-         'If importanize.json is present, that config '
+         'If "{}" is present in either current folder '
+         'or any parent folder, that config '
          'will be used. Otherwise crude default pep8 '
-         'config will be used.{}'.format(found_default),
+         'config will be used.{}'
+         ''.format(IMPORTANIZE_CONFIG,
+                   found_default),
 )
 parser.add_argument(
     '--print',
@@ -113,7 +120,7 @@ def run_importanize(path, config, args):
             return False
 
     text = read(path)
-    artifacts = get_artifacts(path)
+    file_artifacts = get_file_artifacts(path)
 
     lines_iterator = enumerate(iter(text.splitlines()))
     imports = list(parse_statements(find_imports_from_lines(lines_iterator)))
@@ -149,10 +156,10 @@ def run_importanize(path, config, args):
              + lines[first_import_line_number:]
              + [''])
 
-    lines = artifacts.get('sep', '\n').join(lines)
+    lines = file_artifacts.get('sep', '\n').join(lines)
 
     if args.print:
-        print(lines.encode('utf-8'))
+        print(lines.encode('utf-8') if not six.PY3 else lines)
     else:
         with open(path, 'wb') as fid:
             fid.write(lines.encode('utf-8'))